Libft v2

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: rmarceau <rmarceau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/22 13:48:29 by rmarceau          #+#    #+#             */
-/*   Updated: 2022/10/23 13:24:12 by rmarceau         ###   ########.fr       */
+/*   Updated: 2022/10/23 13:31:36 by rmarceau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
